refactor: Altered variables name and implements FileProcess class.
